♻️ [Refactoring]: #041 xref/fallback/ を 1機能1フォルダ化 + xref-pipeline integration test 移動
- xref/fallback/{fallback-scanner,object-scanner} を個別フォルダ化、テストを各 __tests__/ へ
- 主実装内の相対 import を +1、test 内を +2 階層調整
- xref/fallback/index.ts (barrel) は folder/index 解決で無修正
- xref/xref-pipeline.integration.test.ts を xref/__tests__/ へ移動（領域横断 test）
- Test Files=124 / Tests=1688 維持
