+++ packages/core/src/xref/fallback/fallback-scanner/__tests__/fallback-scanner.behavior.test.ts
@@ -3,8 +3,8 @@ import {
   ByteOffset,
   GenerationNumber,
   ObjectNumber,
-} from "../../pdf/types/index";
-import { scanFallback } from "./fallback-scanner";
+} from "../../../../pdf/types/index";
+import { scanFallback } from "../../fallback-scanner";
 
 function encode(s: string): Uint8Array {
   return new TextEncoder().encode(s);
