+++ packages/core/src/xref/fallback/object-scanner/__tests__/object-scanner.boundary.test.ts
@@ -1,5 +1,5 @@
 import { expect, test } from "vitest";
-import { scanObjectHeaders } from "./object-scanner";
+import { scanObjectHeaders } from "../../object-scanner";
 
 function encode(s: string): Uint8Array {
   return new TextEncoder().encode(s);
